+++ src/dragonfly/rules.py
@@ -9,16 +9,21 @@ import aiohttp
 import yara
 
 REPO_ZIP_URL: Final[str] = "https://api.github.com/repos/SkeletalDemise/DragonFly/zipball/"
-HEADERS: Final[dict[str, str]] = {"Authorization": f"Bearer {getenv('DRAGONFLY_GITHUB_TOKEN')}"}
+REPO_TOP_COMMIT_URL: Final[str] = "https://api.github.com/repos/SkeletalDemise/DragonFly/commits/main"
+AUTH_HEADERS: Final[dict[str, str]] = {"Authorization": f"Bearer {getenv('DRAGONFLY_GITHUB_TOKEN')}"}
+JSON_HEADERS: Final[dict[str, str]] = {"Accept": "application/vnd.github.VERSION.sha"}
 
 
-async def _fetch_rules() -> dict[str, str]:
+async def _fetch_rules() -> tuple[str, dict[str, str]]:
     """Return a dictionary mapping filenames to content"""
     files = {}
     buffer = BytesIO()
-    async with aiohttp.ClientSession(raise_for_status=True, headers=HEADERS) as http_session:
+    async with aiohttp.ClientSession(raise_for_status=True, headers=AUTH_HEADERS) as http_session:
         async with http_session.get(REPO_ZIP_URL) as response:
             buffer.write(await response.content.read())
+        async with http_session.get(REPO_TOP_COMMIT_URL, headers=JSON_HEADERS) as response:
+            data = await response.read()
+            rules_commit = data.decode()
     buffer.seek(0)
     with ZipFile(buffer) as zip_file:
         for file_path in zip_file.namelist():
@@ -26,10 +31,10 @@ async def _fetch_rules() -> dict[str, str]:
                 file_name = file_path.split("/")[-1]
                 file_name = file_name.removesuffix(".yara")
                 files[file_name] = zip_file.read(file_path).decode()
-    return files
+    return rules_commit, files
 
 
-async def get_rules() -> "compiled yara rules":
+async def get_rules() -> tuple[str, yara.Rules]:
     """Fetch and compile the rules"""
-    sources = await _fetch_rules()
-    return yara.compile(sources=sources)
+    rules_commit, sources = await _fetch_rules()
+    return rules_commit, yara.compile(sources=sources)